Filter dictionary by word class
