+++ package.json
@@ -29,6 +29,7 @@
     },
     "dependencies": {
         "react-bootstrap": "^1.2.2",
+        "react-device-detect": "^1.13.1",
         "react-flash-message": "^1.0.2",
         "react-infinite-scroller": "^1.2.4",
         "react-ios-pwa-prompt": "^1.8.0",
